+++ functions/upload.py
@@ -3,6 +3,7 @@ from firebase_admin import credentials, firestore, App
 import psp_liquids_daq_parser as daq
 import pandas as pd
 import gorillacompression as gc
+import matplotlib.pyplot as plt
 
 # Use a service account.
 cred: credentials.Certificate = credentials.Certificate(
@@ -20,25 +21,23 @@ db = firestore.client()
 # if test_name == "":
 #     raise RuntimeError
 
-test_name = "cms_sd_hotfire_1_hf"
+test_name = "cms_whoopsie_hf_hf"
 
 parsed_datasets: dict[
     str,
     daq.AnalogChannelData | daq.DigitalChannelData | daq.SensorNetData | list[float],
 ] = daq.parseTDMS(
     5,
-    1713579651000,
-    file_path_custom="C:\\Users\\rajan\\Desktop\\PSP_Data\\sd_hotfire\\DataLog_2024-0419-2120-51_CMS_Data_Wiring_5.tdms",
+    1714536122000,
+    file_path_custom="C:\\Users\\rajan\\Desktop\\PSP_Data\\whoopsie\\DataLog_2024-0501-0002-02_CMS_Data_Wiring_5.tdms",
 )
 parsed_datasets.update(
     daq.parseTDMS(
         6,
-        1713579651000,
-        file_path_custom="C:\\Users\\rajan\\Desktop\\PSP_Data\\sd_hotfire\\DataLog_2024-0419-2120-51_CMS_Data_Wiring_6.tdms",
+        1714536122000,
+        file_path_custom="C:\\Users\\rajan\\Desktop\\PSP_Data\\whoopsie\\DataLog_2024-0501-0002-02_CMS_Data_Wiring_6.tdms",
     )
 )
-
-# sensornet_datasets = daq.parseCSV(1713579651,file_path_custom="C:/Users/rajan/Desktop/PSP_Data/sd_hotfire/reduced_sensornet_data.csv")
 
 dict_to_write: dict[str, list[float]] = {}
 
@@ -46,14 +45,18 @@ all_time: list[float] = parsed_datasets["time"]
 
 available_datasets: list[str] = []
 
+# parsed_datasets = daq.parseCSV(1714536122000, file_path_custom="C:\\Users\\rajan\\Desktop\\PSP_Data\\whoopsie\\reduced_sensornet_data.csv")
+
+
 for dataset in parsed_datasets:
 # dataset = "pt-fu-02"
     if dataset != "time":
         data: list[float] = parsed_datasets[dataset].data.tolist()
+        # time: list[float] = parsed_datasets[dataset].time.tolist()
         time: list[float] = all_time[: len(data)]
         df = pd.DataFrame.from_dict({"time": time, "data": data})
-        df_cut = df.head(20*1000)
-        thing = df_cut.iloc[::10, :]
+        df_cut = df.head(15*1000)
+        thing = df_cut.iloc[::4, :]
         # print("writing csv...")
         # df.to_csv(dataset+".csv", lineterminator="\n",index=False)
         # print("compressing...")
@@ -64,7 +67,7 @@ for dataset in parsed_datasets:
         scale = "psi"
         if "tc" in dataset:
             scale = "deg"
-        if "pi-" in dataset or "reed-" in dataset:
+        if "pi-" in dataset or "reed-" in dataset or "_state" in dataset:
             scale = "bin"
         if "fms" in dataset:
             scale = "lbf"
@@ -82,19 +85,8 @@ for dataset in parsed_datasets:
         )
         available_datasets.append(dataset)
         print(dataset)
-        # print(dataset)
 
-doc_ref = db.collection(test_name).document("general")
-doc_ref.set({"datasets": available_datasets, "test_article": "CMS", "gse_article":"BCLS", "name": "[Autosequence Data] Hotfire 1 (2s)"}, merge=True)
-
-# for sensornet_data in sensornet_datasets:
-#     dict_to_write.update({
-#             sensornet_data + "__time": sensornet_datasets[sensornet_data].time,
-#             sensornet_data: sensornet_datasets[sensornet_data].data
-#         })
+# doc_ref = db.collection(test_name).document("general")
+# doc_ref.set({"datasets": available_datasets, "test_article": "CMS", "gse_article":"BCLS", "name": "[HF:HF] Whoopsie"}, merge=True)
 
 
-# df = pd.DataFrame(dict([ (k,pd.Series(v)) for k,v in dict_to_write.items() ]))
-
-# df.to_csv("compiled.csv", index=False)
-# df.to_json("compiled.json", orient="records", lines=True)